Viz: byte buffers for Beat Particles and Scope Afterglow; setReact
The particle canvas was 128x64x3 float32, 98 KB, and the afterglow 33 KB, both
rewritten every frame at 60 Hz through the S3's 32 KB PSRAM cache: a risk to
the frame rate that cannot be measured without the panel. Both are now bytes
with integer fading and adding, in effects_wow.py and in C++ alike: 24.6 KB and
8.2 KB. What the fractions carried sits below what RGB565 shows. The engine
takes 1,592 B of state and 46,672 B of buffers (host float build), not the
90 KB the previous commit's message said. Engine::setReact applies a new beat
reactivity without resetting the effect.

make -C tools/audiofx/host wow: real = double identical to Python on all eight,
300 frames each; real = float identical on seven, Beat Particles differs in 15
frames by at most 2 px (a particle truncating to the next pixel).

+++ src/viz/wow/wow_internal.h
@@ -121,13 +121,13 @@ struct State {
   int nRings = 0;
 
   // Beat Particles
-  float *rgb = nullptr;           // kH x kW x 3, float32 as RgbCanvas
+  uint8_t *rgb = nullptr;         // kH x kW x 3 bytes, as gfx.RgbCanvas
   Particle *parts = nullptr;
   int nParts = 0;
   Star stars[kStars] = {};
 
   // Scope Afterglow
-  float *glow = nullptr;          // kH x kW, float32
+  uint8_t *glow = nullptr;        // kH x kW bytes, 255 = full
 
   // Twin VU
   real pos[2] = {}, vel[2] = {};
